build: gems balances script
