+++ package.json
@@ -104,13 +104,14 @@
     "start-fork": "FORK_ENABLED=true hardhat node --no-deploy",
     "deploy": "hardhat deploy --write true --network",
     "accounts": "hardhat accounts --network localhost",
+    "gems": "hardhat gems --network localhost",
     "deployments": "hardhat deployments --network localhost",
     "get-some-dai": "hardhat --network localhost fork:get-some-dai",
     "get-some-defo": "hardhat --network localhost get-some-defo",
     "mint-some-gems": "hardhat --network localhost mint-some-gems",
     "jump-in-time": "hardhat --network localhost jump-in-time",
     "permit": "hardhat --network localhost permit",
-    "run-fork": "yarn deploy localhost && yarn get-some-dai && yarn get-some-defo && yarn permit && yarn mint-some-gems && yarn jump-in-time && yarn accounts",
+    "run-fork": "yarn deploy localhost && yarn get-some-dai && yarn get-some-defo && yarn permit && yarn mint-some-gems && yarn jump-in-time && yarn gems && yarn accounts",
     "test": "HIDE_DEPLOY_LOG=true OPTIMIZER_DISABLED=true hardhat unused/test",
     "gas": "REPORT_GAS=true HIDE_DEPLOY_LOG=true hardhat unused/test",
     "verify": "hardhat etherscan-verify --license GPL-3.0 --solc-input --network"
