+++ tasks/gems.ts
@@ -1,41 +1,62 @@
-import chalk from "chalk";
 import { task } from "hardhat/config";
+import _ from "lodash";
 
-import ERC20_ABI from "../abi/defo-abi.json";
-import { announce, info } from "../utils/helpers";
+import { gemName, gems } from "../constants";
+import { ERC721Facet, GemFacet, GemGettersFacet } from "../types";
+import { LibGem } from "../types/contracts/facets/GemGettersFacet";
+import { announce, info, outputFormatKeyValue, outputFormatter, warning } from "../utils/helpers";
 
-task(
-  "accounts",
-  "Get the address and balance information (AVAX, DEFO, DAI, gems) for the accounts.",
-  async (_, hre) => {
-    const { getNamedAccounts, deployments } = hre;
-    const namedAccounts = await getNamedAccounts();
-    const { dai, forkedDefoToken } = namedAccounts;
-    info("\n 📡 Querying balances...");
-    const daiContract = await hre.ethers.getContractAt(ERC20_ABI, dai);
-    const deployedDEFOToken = (await deployments.getOrNull("DEFOToken"))?.address || "";
+export default task("gems", "get gems info and balance information for the deployer").setAction(async (_x, hre) => {
+  const { getNamedAccounts, deployments, ethers } = hre;
+  const { deployer } = await getNamedAccounts();
 
-    const defoContract =
-      forkedDefoToken || deployedDEFOToken
-        ? await hre.ethers.getContractAt("DEFOToken", forkedDefoToken || deployedDEFOToken)
-        : null;
-    announce(
-      `DEFO token is ${chalk.yellow(
-        forkedDefoToken ? "forked" : deployedDEFOToken ? "deployed locally" : chalk.red("not deployed!"),
-      )}. Address: ${defoContract?.address}`,
-    );
+  const diamondDeployment = await deployments.get("DEFODiamond");
+  const gemFacetContract = await ethers.getContractAt<GemFacet>("GemFacet", diamondDeployment.address);
+  const gemGettersFacet = await ethers.getContractAt<GemGettersFacet>("GemGettersFacet", diamondDeployment.address);
+  const gemNFT = await ethers.getContractAt<ERC721Facet>("ERC721Facet", diamondDeployment.address);
+  const types: number[] = Object.values(gems);
 
-    const table = await Promise.all(
-      Object.entries(namedAccounts).map(async ([accountName, accountAddress]) => {
-        return {
-          name: accountName,
-          address: accountAddress,
-          AVAX: Number(hre.ethers.utils.formatEther(await hre.ethers.provider.getBalance(accountAddress))),
-          DAI: Number(hre.ethers.utils.formatEther(await daiContract.balanceOf(accountAddress))),
-          DEFO: defoContract && Number(hre.ethers.utils.formatEther(await defoContract.balanceOf(accountAddress))),
-        };
-      }),
-    );
-    console.table(table);
-  },
-);
+  const gemIds = await gemFacetContract.getGemIdsOf(deployer);
+  const gemsIdsWithData = await Promise.all(
+    gemIds.map(async gemId => {
+      return { gemId: Number(gemId), ...(await gemGettersFacet.GemOf(gemId)) };
+    }),
+  );
+
+  const gemsGroupedByType = gemsIdsWithData.reduce(
+    (r, v, i, a, k = v.GemType) => ((r[k] || (r[k] = [])).push(v), r),
+    {} as Array<Array<LibGem.GemStructOutput & { gemId: number }>>,
+  );
+  announce(`Deployer ${deployer} has ${await gemNFT.balanceOf(deployer)} gem(s)`);
+  info(`Total Charity: ${await gemGettersFacet.getTotalCharity()}`);
+
+  for (const type of types) {
+    warning(`\n\nGem ${gemName(type)} (type ${type})`);
+    announce("Gem config:");
+    console.table([
+      outputFormatter<LibGem.GemTypeMetadataStruct>(
+        ["LastMint", "MaintenanceFee", "RewardRate", "DailyLimit", "MintCount", "DefoPrice", "StablePrice"],
+        await gemGettersFacet.GetGemTypeMetadata(type),
+      ),
+    ]);
+
+    announce(`User balance (${gemsGroupedByType[type].length}):`);
+    const userGems = gemsGroupedByType[type].map(gem => {
+      const pickedGem = _.pick(gem, [
+        "gemId",
+        "MintTime",
+        "LastReward",
+        "LastMaintained",
+        "TaperCount",
+        "booster",
+        "claimedReward",
+      ]) as unknown as Record<string, number | string>;
+      const formattedGem: Record<string, string | number> = {};
+      Object.keys(pickedGem).map(function (key) {
+        formattedGem[key] = outputFormatKeyValue(key, pickedGem[key]);
+      });
+      return formattedGem;
+    });
+    console.table(userGems);
+  }
+});
