+++ tasks/index.ts
@@ -2,6 +2,7 @@ export * as accounts from "./accounts";
 export * as deployments from "./deployments";
 export * as jumpInTime from "./jumpInTime";
 export * as mintGems from "./mintGems";
+export * as gems from "./gems";
 export * as getSomeDefo from "./getSomeDefo";
 export * as permit from "./permit";
 // export * as permit from "./permit";
